+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/MiniAppDownloaderSpec.kt
@@ -294,12 +294,13 @@ class MiniAppDownloaderSpec {
         runBlockingTest {
             val metadataEntity = MetadataEntity(
                 MetadataResponse(
-                    listOf(requiredPermissionObj), listOf(optionalPermissionObj), hashMapOf()
+                    listOf(requiredPermissionObj), listOf(optionalPermissionObj), TEST_ATP_LIST, hashMapOf()
                 )
             )
             val apiManifest = MiniAppManifest(
                 listOf(Pair(MiniAppCustomPermissionType.USER_NAME, "")),
                 listOf(Pair(MiniAppCustomPermissionType.PROFILE_PHOTO, "")),
+                TEST_ATP_LIST,
                 emptyMap()
             )
 
@@ -319,12 +320,16 @@ class MiniAppDownloaderSpec {
         runBlockingTest {
             val metadataEntity = MetadataEntity(
                 MetadataResponse(
-                    listOf(requiredPermissionObj), listOf(optionalPermissionObj), hashMapOf()
+                    listOf(requiredPermissionObj),
+                    listOf(optionalPermissionObj),
+                    TEST_ATP_LIST,
+                    hashMapOf()
                 )
             )
             val cachedManifest = MiniAppManifest(
                 listOf(Pair(MiniAppCustomPermissionType.USER_NAME, "")),
                 listOf(Pair(MiniAppCustomPermissionType.PROFILE_PHOTO, "")),
+                TEST_ATP_LIST,
                 emptyMap()
             )
 
@@ -343,7 +348,10 @@ class MiniAppDownloaderSpec {
         runBlockingTest {
             val metadataEntity = MetadataEntity(
                 MetadataResponse(
-                    listOf(requiredPermissionObj), listOf(optionalPermissionObj), hashMapOf()
+                    listOf(requiredPermissionObj),
+                    listOf(optionalPermissionObj),
+                    TEST_ATP_LIST,
+                    hashMapOf()
                 )
             )
 
@@ -368,7 +376,8 @@ class MiniAppDownloaderSpec {
         runBlockingTest {
             val metadataEntity = MetadataEntity(
                 MetadataResponse(
-                    listOf(requiredPermissionObj), listOf(optionalPermissionObj), hashMapOf()
+                    listOf(requiredPermissionObj), listOf(optionalPermissionObj),
+                    TEST_ATP_LIST, hashMapOf()
                 )
             )
             val requiredPermission =
@@ -380,12 +389,11 @@ class MiniAppDownloaderSpec {
             When calling downloader.listOfPermissions(listOf(optionalPermissionObj)) itReturns optionalPermission
 
             val actual = downloader.prepareMiniAppManifest(metadataEntity)
-
             val requiredPermissions =
                 listOf(Pair(MiniAppCustomPermissionType.USER_NAME, "reason for user name"))
             val optionalPermissions =
                 listOf(Pair(MiniAppCustomPermissionType.PROFILE_PHOTO, "reason for profile photo"))
-            val expected = MiniAppManifest(requiredPermissions, optionalPermissions, hashMapOf())
+            val expected = MiniAppManifest(requiredPermissions, optionalPermissions, TEST_ATP_LIST, hashMapOf())
 
             assertEquals(expected, actual)
         }
@@ -393,7 +401,8 @@ class MiniAppDownloaderSpec {
     @Test
     fun `prepareMiniAppManifest should return empty values correctly`() =
         runBlockingTest {
-            val metadataEntity = MetadataEntity(MetadataResponse(null, null, null))
+            val metadataEntity = MetadataEntity(MetadataResponse(
+                null, null, emptyList(), null))
 
             When calling apiClient.fetchMiniAppManifest(
                 TEST_ID_MINIAPP,
@@ -401,7 +410,7 @@ class MiniAppDownloaderSpec {
             ) itReturns metadataEntity
 
             val actual = downloader.prepareMiniAppManifest(metadataEntity)
-            val expected = MiniAppManifest(emptyList(), emptyList(), emptyMap())
+            val expected = MiniAppManifest(emptyList(), emptyList(), emptyList(), emptyMap())
 
             assertEquals(expected, actual)
         }